Now clang-formatted.

+++ Qt/sightlines/sightlines.c
@@ -4,21 +4,17 @@
 static double EARTH_RADIUS = 6371.0;
 
 /* takes elevation in rads and altitude in kms, returns -1.0
-   on invalid inputs. Elevation must be between 0.01 and 
+   on invalid inputs. Elevation must be between 0.01 and
    pi/2 - 0.01*, altitude between 100 and 50000. */
 #ifdef __cplusplus
 extern "C" {
 #endif
-long double get_sightline(const double ELEVATION, const double ALTITUDE)
+long double get_sightline(const long double ELEVATION, const long double ALTITUDE)
 {
-    if (ELEVATION < 0.01 || ELEVATION > 1.5607963267948965 ||
-        ALTITUDE < 100.0 || ALTITUDE > 50000.0)
+    if (ELEVATION < 0.01 || ELEVATION > 1.5607963267948965 || ALTITUDE < 100.0 || ALTITUDE > 50000.0)
         return -1.0;
-    
-    const long double NUMERATOR = (EARTH_RADIUS + ALTITUDE) * 
-        cos(ELEVATION + 
-        asin(cos(ELEVATION) * (EARTH_RADIUS / 
-        (EARTH_RADIUS + ALTITUDE))));
+
+    const long double NUMERATOR = (EARTH_RADIUS + ALTITUDE) * cos(ELEVATION + asin(cos(ELEVATION) * (EARTH_RADIUS / (EARTH_RADIUS + ALTITUDE))));
     const long double DENOMINATOR = cos(ELEVATION);
     const long double SIGHTLINE = (NUMERATOR / DENOMINATOR);
     return SIGHTLINE;
